+++ frontend/src/components/InfoTooltip.tsx
@@ -1,18 +1,31 @@
+import clsx from "clsx";
+
+import dynamic from "next/dynamic";
 import { forwardRef, ReactNode } from "react";
 
-import { TooltipWrapper } from "./TooltipWrapper";
 import { USWDSIcon } from "./USWDSIcon";
 
 export interface InfoTooltipProps {
   text: ReactNode;
   position?: "top" | "bottom" | "left" | "right";
   className?: string;
+  wrapperClasses?: string;
+  title?: string;
 }
+
+const DynamicTooltipWrapper = dynamic(
+  () => import("src/components/TooltipWrapper"),
+  {
+    ssr: false, // works around bug with Trussworks assigning different random ids on server and client renders
+  },
+);
 
 const InfoTooltip = ({
   text,
   position = "top",
   className,
+  wrapperClasses,
+  title,
 }: InfoTooltipProps) => {
   const IconWrapper = forwardRef<
     HTMLSpanElement,
@@ -22,7 +35,7 @@ const InfoTooltip = ({
       {...props}
       ref={ref}
       style={{ cursor: "help" }}
-      className="text-secondary"
+      className={clsx("text-secondary", props.className)}
     >
       <USWDSIcon name="info_outline" />
     </span>
@@ -31,15 +44,17 @@ const InfoTooltip = ({
   IconWrapper.displayName = "IconWrapper";
 
   return (
-    <TooltipWrapper
+    <DynamicTooltipWrapper
+      title={title}
       label={text}
       position={position}
       asCustom={IconWrapper}
       className={className}
+      wrapperclasses={wrapperClasses}
       data-testid="tooltipWrapper"
     >
       <></>
-    </TooltipWrapper>
+    </DynamicTooltipWrapper>
   );
 };
